% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_crossfit.R
\name{do_crossfit}
\alias{do_crossfit}
\title{do_crossfit}
\usage{
do_crossfit(y, a, x, outfam = gaussian(), treatfam = binomial(),
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.mean", "SL.ranger",
  "SL.glm.interaction"), nsplits = 5)
}
\arguments{
\item{y}{nx1 outcome vector in [0, 1]}

\item{a}{nx1 treatment received vector}

\item{x}{nxp matrix of covariates}

\item{outfam}{family specifying the error distribution for outcome 
regression, currently gaussian() or binomial(), should not specify link.}

\item{treatfam}{family specifying the error distribution for treatment 
regression, currently gaussian() or binomial(), should not specify link.}

\item{sl.lib}{vector specifying which libraries to use for the SL.}

\item{nsplits}{number of splits for the cross-fitting.}
}
\description{
do_crossfit estimates the nuisance regression functions via cross-fitting.
This allows the user to avoid imposing empirical process conditions on these
functions, while still attaining fast rates of convergence.
}
