% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bound.R
\name{get_bound}
\alias{get_bound}
\title{get_bound}
\usage{
get_bound(y, a, x, out_model, treat_model, eps = 0, delta = 0,
  nsplits = 5, do_mult_boot = TRUE)
}
\arguments{
\item{y}{nx1 outcome vector in [0, 1]}

\item{a}{nx1 treatment received vector}

\item{x}{nxp matrix of covariates}

\item{out_model}{a string specifying the outcome model for E(Y|A=a, X)
currently only "logistic" and "ranger" supported.}

\item{treat_model}{a string specifying the outcome model for P(A=1| X)
currently only "logistic" and "ranger" supported.}

\item{eps}{vector of arbitrary length specifying the values for the 
proportion of confounding where the lower and upper bounds curves are 
evaluated.}

\item{delta}{vector of arbitrary length specifying the maximum bias allowed
for the confounded units.}

\item{nsplits}{number of splits for the cross-fitting (default=5)}

\item{do_mult_boot}{boolean for whether uniform bands via the multiplier 
bootstrap need to computed (default is do_mult_boot=TRUE)}
}
\description{
get_bound is the main function to estimate the lower and upper bounds curves.
}
