% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ci}
\alias{get_ci}
\title{get_ci}
\usage{
get_ci(muhat, sigmahat, c)
}
\arguments{
\item{muhat}{the n x 1 vector of means}

\item{sigmahat}{the variance}

\item{c}{the multiplier (e.g. 1.96)}
}
\description{
get_ci constructs a Wald-type confidence interval
}
