% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps_zero.R
\name{get_eps_zero}
\alias{get_eps_zero}
\title{Estimation of the minimum proportion of confounded units such that
the bounds on the average treatment effect contain zero (eps0)}
\usage{
get_eps_zero(n, eps, pt_est, lb, ub, ql, qu, ifvals_lb, ifvals_ub,
  delta = 1, alpha = 0.05)
}
\arguments{
\item{n}{sample size.}

\item{eps}{vector of arbitrary length specifying the values for the 
proportion of confounding where the lower and upper bounds curves are 
evaluated.}

\item{pt_est}{point-estimate under no unmeasured confounding. Generally,
\code{mean(nuhat)}, where \code{nuhat} is a vector of influence fn values
for the parameter E{E(Y|A = 1, X) - E(Y|A = 0, X)}.}

\item{lb}{length(eps) x (num of delta values) matrix containing the values 
for the lower bounds.}

\item{ub}{length(eps) x (num of delta values) matrix containing the values 
for the upper bounds.}

\item{ql}{a list of size \code{nsplits}, where element j is a
(num eps) x (num delta) matrix containing estimates of eps-quantile of 
g(etab) for lower bound computed using all obs except those in fold j.}

\item{qu}{a list of size \code{nsplits}, where element j is a
(num eps) x (num delta) matrix containing estimates of eps-quantile of 
g(etab) for upper bound computed using all obs except those in fold j.}

\item{ifvals_lb}{a list of size \code{nsplits}, where element j
is n x length(eps) x length(delta) array containing
values for \code{ifvals_lb} - \code{lambda_lb} * \code{q_lb} computed using
regression functions estimated from all obs except those in fold j and 
evaluated at obs in fold j.}

\item{ifvals_ub}{a list of size \code{nsplits}, where element j
is n x length(eps) x length(delta) array containing
values for \code{ifvals_ub} - \code{lambda_ub} * \code{q_ub} computed using
regression functions estimated from all obs except those in fold j and 
evaluated at obs in fold j.}

\item{delta}{vector of delta values specifying the values of delta used
to bound maximal confounding among S = 0 units. Length(\code{delta}) should 
match the 2nd dimension of \code{lb}, \code{ub}, \code{ql}, \code{qu} and the
3rd dimension of \code{ifvals_lb} and \code{ifvals_ub}.}

\item{alpha}{scalar specifying the confidence level. Default is 0.05.}
}
\value{
A length(delta)x5 \code{data.frame} with values of delta, estimate
of eps0, \code{max}(0, ci_lo), \code{min}(1, ci_hi), variance of estimate of 
eps0.
}
\description{
\code{get_eps_zero} computes an estimate of eps0 and Wald-type
confidence interval.
}
\seealso{
\code{\link{if_gamma}}, \code{\link{if_tau}},  
\code{\link{get_ifvals}}.
}
