% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_g.R
\name{get_g}
\alias{get_g}
\title{get_g}
\usage{
get_g(pi0, pi1, mu0, mu1, upper = FALSE, max_bias = 1)
}
\arguments{
\item{pi0}{nx1 vector with estimates for P(A=0 | X=x) evaluated at each x}

\item{pi1}{nx1 vector with estimates for P(A=1 | X=x) evaluated at each x}

\item{mu0}{nx1 vector with estimates for E(Y |A=0, X=x) evaluated at each x}

\item{mu1}{nx1 vector with estimates for E(Y |A=1, X=x) evaluated at each x}

\item{upper}{boolean (default is upper=FALSE) if TRUE g_u(eta) is returned}

\item{max_bias}{scalar specifying maximum bias for confounded units 
(default is max_bias=1)}
}
\description{
get_g is the function that computes g_l(upper=FALSE) or g_u (upper=TRUE) 
given estimates of the nuisance regression functions evaluated at each value 
of X and a value for max_bias (default=1) specifyin the max_bias allowed for 
the confounded units.
}
