% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_im04.R
\name{get_im04}
\alias{get_im04}
\title{get_im04}
\usage{
get_im04(n, lb, ub, sigma_l, sigma_u, alpha)
}
\arguments{
\item{n}{number of observations}

\item{lb}{estimate of lower bound}

\item{ub}{estimate of upper bound}

\item{sigma_l}{estimate of standard deviation of lb}

\item{sigma_u}{estimate of standard deviation of ub}

\item{alpha}{desired significance level}
}
\description{
get_im04 computes c_alpha as in Imbens and Manski (2004)
}
