% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muahat.R
\name{get_muahat}
\alias{get_muahat}
\title{get_muahat}
\usage{
get_muahat(y, a, x, newx, aval, method = "logistic", trunc_tol = 0)
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{x}{nxp matrix of covariates}

\item{newx}{qxp matrix of values that the regression estimates are evaluated at}

\item{aval}{scalar specifying which regression to estimate E(Y|A=aval,X=x)}

\item{method}{a string specifying the method to use for estimation. Currently
supported is "logistic" and "ranger" (for random forest with ranger package)}

\item{trunc_tol}{scalar specifying the amount that the predicted values need 
to be truncated by. Yhat is in [trunc_tol, 1-trunc_tol]. Default is 0.}
}
\description{
get_muahat estimates the regression function E(Y|A=aval,X=x) and evaluate it
at newx.
}
