% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regressions.R
\name{get_piahat}
\alias{get_piahat}
\title{Estimation of regression function E(A|X)}
\usage{
get_piahat(a, x, newx, family = binomial(), trunc_tol = 0.05,
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.mean",
  "SL.glm.interaction", "SL.ranger"))
}
\arguments{
\item{a}{nx1 vector of treatments}

\item{x}{nxp \code{data.frame} of covariates}

\item{newx}{qxp \code{data.frame} of values that the regression estimates are 
evaluated at}

\item{family}{family specifying the error distribution for treatment
regression, currently only \code{binomial()} supported. 
Link should not be specified.}

\item{trunc_tol}{amount of tolerance allowed for truncating the propensity 
score in [tol, 1 - tol]. Default is 0.05.}

\item{sl.lib}{character vector specifying which libraries to use for the SL.}
}
\value{
A n-dimensional vectors containing estimates of E(Y|A = aval, X) 
evaluated at newx.
}
\description{
\code{get_piahat} estimates the regression function E(A|X)
using the SuperLearner.
}
\section{Details}{
 
If the SuperLearner returns an error, a GLM is fitted instead. In this
case the user is suggested to choose some other method of estimation and then
pass the estimates as arguments to other functions. Sometimes SuperLearner
returns warnings messages.
}

\examples{
n <- 500
x <- data.frame(x1 = rnorm(n), x2 = runif(n))
newx <- data.frame(x1 = rnorm(20), x2 = runif(20))
a <- rbinom(n, 1, pnorm(x$x1))
fits <- get_piahat(a, x, newx, family = binomial(), 
                   sl.lib = c("SL.mean", "SL.glm", "SL.gam"))
head(fits)

}
\references{
Van der Laan, M. J., Polley, E. C., & Hubbard, A. E. (2007). 
Super learner. 
\emph{Statistical applications in genetics and molecular biology}, 6(1).
}
\seealso{
\code{\link{do_crossfit}} and \code{\link{get_muahat}}.
}
