% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_gamma.R
\name{if_gamma}
\alias{if_gamma}
\title{if_gamma}
\usage{
if_gamma(y, a, aval, pia, mua)
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{aval}{scalar, if aval=a then it returns the if vals for E(E(Y|A=a, X)))}

\item{pia}{nx1 vector with estimates for P(A=aval | X) evaluated at each X.}

\item{mua}{nx1 vector with estimates for E(Y | A=aval, X) evaluated at each 
X.}
}
\description{
if_gamma returns the influence function values for the parameter 
E(E(Y|A=a, X))) evaluated at each X.
}
