% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(simmat, truth, n)
}
\arguments{
\item{simmat}{a n x nsim matrix, where n is the number of epsilon points used
in evaluation and nsim is the number of simulations}

\item{truth}{a n x 1 vector containing the true values for the curve}

\item{n}{is the sample size}
}
\description{
rmse computes RMSE, see Section 5 in paper
}
