% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivitypuc.R
\docType{package}
\name{sensAteBounds}
\alias{sensAteBounds}
\alias{sensAteBounds-package}
\title{echseffects: A package to perform sensitivity analysis via the proportion of 
unmeasured confounding.}
\description{
echseffects: A package to perform sensitivity analysis via the proportion of 
unmeasured confounding.
}
