% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_by}
\alias{summarize_by}
\title{summarize_by
Wrapper for group_by + summarize_all from dplyr}
\usage{
summarize_by(x, fun, ...)
}
\arguments{
\item{x}{a matrix}

\item{fun}{a function (as a string) that we want to apply to the values in 
the matrix}

\item{...}{some columns in x which we want to apply fun conditional on}
}
\description{
summarize_by
Wrapper for group_by + summarize_all from dplyr
}
