% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bias}
\alias{bias}
\title{bias}
\usage{
bias(simmat, truth)
}
\arguments{
\item{simmat}{a n x nsim matrix, where n is the number of epsilon points used
in evaluation and nsim is the number of simulations. If it is a nsim-dim 
vector, the length is interpreted as the number of simulations of a scalar
quantity.}

\item{truth}{a n x 1 vector containing the true values for the curve.
If simmat is a vector, truth must be scalar.}
}
\description{
bias computes the bias, see Section 5 for definition
}
