% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_crossfit.R
\name{do_crossfit}
\alias{do_crossfit}
\title{do_crossfit}
\usage{
do_crossfit(y, a, x, ymin, ymax, outfam = gaussian(),
  treatfam = binomial(), nsplits = 5, do_parallel = FALSE,
  ncluster = NULL, sl.lib = c("SL.earth", "SL.gam", "SL.glm",
  "SL.mean", "SL.ranger", "SL.glm.interaction"))
}
\arguments{
\item{y}{nx1 outcome vector in [0, 1]}

\item{a}{nx1 treatment received vector}

\item{x}{nxp matrix of covariates}

\item{outfam}{family specifying the error distribution for outcome 
regression, currently gaussian() or binomial(), should not specify link.}

\item{treatfam}{family specifying the error distribution for treatment 
regression, currently gaussian() or binomial(), should not specify link.}

\item{nsplits}{number of splits for the cross-fitting.}

\item{do_parallel}{boolean for whether parallel computing should be used}

\item{ncluster}{number of clusters used if parallel computing is used.}

\item{sl.lib}{vector specifying which libraries to use for the SL.}
}
\description{
do_crossfit estimates the nuisance regression functions via cross-fitting.
This allows the user to avoid imposing empirical process conditions on these
functions, while still attaining fast rates of convergence.
}
