% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_multboot.R
\name{do_multboot}
\alias{do_multboot}
\title{Multiplier bootstrap to construct uniform confidence bands around
the curves tracing the bounds on ATE.}
\usage{
do_multboot(n, psihat, sigmahat, ifvals, alpha = 0.05, B = 10000)
}
\arguments{
\item{n}{sample size.}

\item{psihat}{Estimate of either lower or upper bound curve minus 
(epsilon x an estimate of the appropriate quantile). It is a vector 
of length equal to the number of epsilon points at which the lower and upper 
bounds curves are evaluated.}

\item{sigmahat}{Estimate of the variance of psihat at each value of eps grid.
It is a vector of length equal to the number of epsilon points at which the 
lower and upper bounds curves are evaluated.}

\item{ifvals}{n times neps matrix, where n is number of observations, and 
neps is the number of epsilon points at which the lower and upper bounds 
curves are evaluated. It represents the "influence function" values 
associated with psihat.}

\item{alpha}{confidence level. Default is 0.05.}

\item{B}{number of rademacher rvs sampled. Default is 10000.}
}
\value{
a scalar calpha equal to the multiplier used to construct bands of 
the form psi(eps) \pm calpha * sigma(eps).
}
\description{
\code{do_multboot} is the main function to implement 
the multiplier bootstrap to estimate uniformly valid bands around the bounds.
}
\references{
Kennedy, E. H. (2019). Nonparametric causal effects based on 
incremental propensity score interventions. \emph{Journal of the American 
Statistical Association}, 114(526), 645-656.
}
