% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ci}
\alias{get_ci}
\title{get_ci}
\usage{
get_ci(muhat, sigmahat, c)
}
\arguments{
\item{muhat}{the n x 1 vector of means}

\item{sigmahat}{the nx1 vector of variances}

\item{c}{the multiplier (e.g. 1.96)}
}
\value{
a nx2 matrix containing lower and upper endpoints of confidence
interval.
}
\description{
\code{get_ci} constructs a Wald-type confidence interval
}
