% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esp_zero.R
\name{get_eps_zero}
\alias{get_eps_zero}
\title{get_esp_zero}
\usage{
get_eps_zero(n, eps, lb, ub, ql, qu, ifvals_lb, ifvals_ub, delta,
  alpha = 0.05)
}
\arguments{
\item{eps}{vector of arbitrary length specifying the values for the 
proportion of confounding where the lower and upper bounds curves are 
evaluated.}

\item{lb}{vector of length(eps) containing the values for the lower bound}

\item{ub}{vector of length(eps) containing the values for the upper bound}

\item{ql}{vector of length(eps) containing the eps quantiles of g_l(eta)}

\item{qu}{vector of length(eps) containing the eps quantiles of g_u(eta)}

\item{ifvals_lb}{matrix nxlength(eps) containg the if values for lb:
lb - ql * lambdal, where lambdal is the indicator for g_l(eta) <= ql}

\item{ifvals_ub}{matrix nxlength(eps) containg the if values for ub:
ub - qu * lambdau, where lambdau is the indicator for g_u(eta) > qu}

\item{alpha}{scalar specifying the confidence level (default is alpha=0.05)}
}
\description{
get_eps_zero computes an estimate of esp_zero and an associated 
confidence interval based on Theorem 4 in manuscript.
}
