% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_functions.R
\name{get_g}
\alias{get_g}
\title{Estimation of the function g(eta)}
\usage{
get_g(data, delta, upper)
}
\arguments{
\item{data}{data.frame of dimensions n x 6, with columns: ymin, ymax, 
pi0g (= P(A = 0 | X = x) if model "x" or 1-A if model "xa"),
pi1g (= P(A = 1 | X = x) if model "x" or A if model "xa")
mu1 (E(Y | A = 1, X = x)) and mu0 (E(Y | A = 0, X = x)).}

\item{delta}{vector of values for the delta parameter.}

\item{upper}{boolean (default is upper=FALSE) if TRUE values for 
g_u(eta) are returned, otherwise those for g_l(eta) are returned.}
}
\value{
a nrow(data) x ndelta matrix containing the values of g(eta) at 
each observed X.
}
\description{
\code{get_g} returns the values for the g_l(eta) 
(upper=FALSE) or g_u(eta) (upper=TRUE) at each X.
}
\section{Details}{

As done in the paper, one can see that g(eta) for the lower bound is equal to
g(eta) for the upper bound minus delta * (ymax - ymin). 
They are keep separated just for code clarity.
}

\seealso{
\code{\link{if_gamma}}, \code{\link{get_ifvals}}, 
\code{\link{if_tau}}.
}
