% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_functions.R
\name{get_ifvals}
\alias{get_ifvals}
\title{Estimation of influence functions for the bounds on ATE}
\usage{
get_ifvals(n, neps, ndelta, upper, nu, tau, g, nsplits, quant, nobs_fold)
}
\arguments{
\item{n}{sample size.}

\item{upper}{boolean for whether the upper or the lower bound curve needs to
be computed.}

\item{nu}{a list of size \code{nsplits}, where element j 
is a (num obs in split j)x1 matrix containing the influence function values 
for E{E(Y|A = 1, X) - E(Y|A = 0, X)} computed using regressions fns estimated 
using all obs except those in fold j and evaluated at obs in fold j.}

\item{tau}{a list of size \code{nsplits}, where element j 
is a (num obs in split j)x1 matrix containing the influence function values 
for E{g(eta)} computed using regressions fns estimated 
using all obs except those in fold j and evaluated at obs in fold j.}

\item{eps}{vector of arbitrary length specifying the values for the 
proportion of confounding where the lower and upper bounds curves are 
evaluated.}

\item{ghatmat}{a list of size \code{nsplits}, where element j 
is a (num obs in split j)xlength(delta) matrix containing the 
values of g(eta) for lower bound if \code{upper = FALSE}, for the upper
bound otherwise, computed using regressions fns estimated 
using all obs except those in fold j and evaluated at obs in fold j.}

\item{delta}{vector of delta values. \code{length}(delta) should match 
\code{ncol(ghatmat)}.}
}
\description{
\code{get_ifvals} is the main function for getting the influence 
function values used to construct the estimator of either the lower and upper 
bound curves.
}
\seealso{
\code{\link{if_gamma}}, \code{\link{if_tau}}.
}
