% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ifvals.R
\name{get_ifvals}
\alias{get_ifvals}
\title{get_ifvals}
\usage{
get_ifvals(eps, upper, nu, tau, ghatmat)
}
\arguments{
\item{eps}{vector of arbitrary length specifying the values for the 
proportion of confounding where the lower and upper bounds curves are 
evaluated.}

\item{upper}{boolean for whether the upper or the lower bound curve needs to
be computed.}

\item{nu}{nx1 vector of influence function values for the parameter 
E(E(Y|A = 1, X) - E(Y|A = 0, X))}

\item{tau}{nx1 vector of influence function values for the parameter
E(g_l(eta)) is upper=FALSE or E(g_u(eta)) if upper=TRUE. See manuscript.}

\item{ghatmat}{nxd matrix containing the values of g_l(eta)
if upper=FALSE or g_u(eta) if upper=TRUE, evaluated at each values of X. 
Each column of the matrix represents the vector of values for a specific 
value of delta, the maximum bias allowed for the confounded units. See 
manuscript.}
}
\description{
get_ifvals is the main function for getting the influence function values
used to construct the estimator of either the lower or the upper bound curve.
}
