% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imbens_maski_2004.R
\name{get_im04}
\alias{get_im04}
\title{Estimation of multiplier for confidence interval for ATE as described
in Imbens and Manski (2004)}
\usage{
get_im04(dat, n, alpha)
}
\arguments{
\item{dat}{vector containing estimate of lower bound, estimate of upper bound,
estimate of standard deviation of lb, estimate of standard deviation of ub in
this order.}

\item{n}{sample size.}

\item{alpha}{desired significance level (capped at \code{pnorm(5)}).}
}
\value{
a scalar equal to the multiplier used to construct the confidence 
interval for partially identified ATE.
}
\description{
\code{get_im04} computes c_alpha as in Imbens and Manski (2004).
}
\examples{
n <- 100
dat <- c(-0.4, 0.3, 0.2, 0.1)
get_im04(dat, n, 0.05) # large identification interval width
dat2 <- c(-0.4, -0.395, 0.2, 0.1)
get_im04(dat2, n, 0.05) # small identification interval width

}
\references{
Imbens, G. W., & Manski, C. F. (2004). Confidence intervals for 
partially identified parameters. \emph{Econometrica}, 72(6), 1845-1857.
}
