% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_regressions.R
\name{get_muahat}
\alias{get_muahat}
\title{Estimation of regression function E(Y|A = a, X)}
\usage{
get_muahat(y, a, x, newx, aval, ymin, ymax, family = gaussian(),
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.ranger",
  "SL.glm.interaction", "SL.mean"))
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{x}{nxp \code{data.frame} of covariates}

\item{newx}{qxp \code{data.frame} of values that the regression estimates are 
evaluated at}

\item{aval}{Scalar specifying which regression to estimate E(Y|A=aval,X=x)}

\item{ymin}{Infimum of the support of y, so that predictions will be forced 
to be >= ymin.}

\item{ymax}{Supremum of the support of y, so that predictions will be forced 
to be <= ymax.}

\item{family}{family specifying the error distribution for outcome 
regression, currently \code{gaussian()} or \code{binomial()} supported. 
Link should not be specified. Default is \code{gaussian()}.}

\item{sl.lib}{character vector specifying which libraries to use for the SL.}
}
\value{
A n-dimensional vectors containing estimates of E(Y|A = aval, X) 
evaluated at newx.
}
\description{
\code{get_muahat} estimates the regression function 
E(Y|A = a, X) using the SuperLearner.
}
\section{Details}{
 
If the SuperLearner returns an error, a GLM is fitted instead. In this
case the user is suggested to choose some other method of estimation and then
pass the estimates as arguments to other functions. Sometimes SuperLearner
returns warnings messages.
}

\examples{
n <- 500
x <- data.frame(x1 = rnorm(n), x2 = runif(n))
newx <- data.frame(x1 = rnorm(20), x2 = runif(20))
a <- rbinom(n, 1, pnorm(x$x1))
y <- 2 + x$x1 - x$x2 + rnorm(n)
fits <- get_muahat(y, a, x, newx, 1, min(y), max(y), family = gaussian(), 
                   sl.lib = c("SL.mean", "SL.glm", "SL.gam"))
head(fits)

}
\references{
Van der Laan, M. J., Polley, E. C., & Hubbard, A. E. (2007). 
Super learner. 
\emph{Statistical applications in genetics and molecular biology}, 6(1).
}
\seealso{
\code{\link{do_crossfit}} and \code{\link{get_piahat}}.
}
