% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muahat.R
\name{get_muahat}
\alias{get_muahat}
\title{get_muahat}
\usage{
get_muahat(y, a, x, newx, aval, trunc_tol = 0, family = gaussian(),
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.glm.interaction",
  "SL.mean", "SL.ranger"))
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{x}{nxp matrix of covariates}

\item{newx}{qxp matrix of values that the regression estimates are evaluated at}

\item{aval}{Scalar specifying which regression to estimate E(Y|A=aval,X=x)}

\item{trunc_tol}{Scalar specifying the amount that the predicted values need 
to be truncated by. Yhat is in [trunc_tol, 1-trunc_tol]. Default is 0.}

\item{family}{Currently allows gaussian or binomial to describe the error 
distribution. It is passed as argument to SL fun, should not contain link.}
}
\description{
get_muahat estimates the regression function E(Y|A=aval,X=x) and evaluate it
at newx.
}
