% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muahat.R
\name{get_muahat}
\alias{get_muahat}
\title{get_muahat}
\usage{
get_muahat(y, a, x, newx, aval, ymin, ymax, family = gaussian(),
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.ranger",
  "SL.glm.interaction", "SL.mean"))
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{x}{nxp matrix of covariates}

\item{newx}{qxp matrix of values that the regression estimates are evaluated at}

\item{aval}{Scalar specifying which regression to estimate E(Y|A=aval,X=x)}

\item{ymin}{infimum of the support of y (predictions will be >= ymin)}

\item{ymax}{supremum of the support of y (predictions will be <= ymax)}

\item{family}{Currently allows gaussian or binomial to describe the error 
distribution. It is passed as argument to SL fun, should not contain link.}
}
\description{
get_muahat estimates the regression function E(Y|A=aval,X=x) and evaluate it
at newx.
}
