% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multboot.R
\name{get_multboot}
\alias{get_multboot}
\title{get_multboot}
\usage{
get_multboot(n, psihat, sigmahat, ifvals, alpha = 0.05)
}
\arguments{
\item{n}{the number of observations}

\item{psihat}{ADD}

\item{sigmahat}{ADD}

\item{ifvals}{ADD}

\item{alpha}{confidence level (default is 0.05)}
}
\description{
get_multboot is the main function to implement the multiplier bootstrap
}
