% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_piahat.R
\name{get_piahat}
\alias{get_piahat}
\title{get_piahat}
\usage{
get_piahat(a, x, newx, family = binomial(), trunc_tol = 0.05,
  sl.lib = c("SL.earth", "SL.gam", "SL.glm", "SL.glm.interaction",
  "SL.mean", "SL.ranger"))
}
\arguments{
\item{a}{nx1 vector of treatments}

\item{x}{nxp matrix of covariates}

\item{newx}{qxp matrix of values that the regression estimates are evaluated at}

\item{family}{Currently allows gaussian or binomial to describe the error 
distribution. It is passed as argument to SL fun, should not contain link.}

\item{trunc_tol}{scalar specifying the amount that the predicted values need 
to be truncated by. pihat is in [trunc_tol, 1-trunc_tol]. Default is 0.05.}
}
\description{
get_piahat estimates the regression function P(A=1|X=x) and evaluate it
at newx.
}
