% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_functions.R
\name{if_tau}
\alias{if_tau}
\title{Estimation of influence function for E(g(eta))}
\usage{
if_tau(y, a, ymin, ymax, pi0, pi1, mu0, mu1, upper = FALSE)
}
\arguments{
\item{y}{nx1 vector of outcomes in [0, 1]}

\item{a}{nx1 vector of treatments}

\item{pi0}{nx1 vector with estimates for P(A=0 | X=x) evaluated at each x}

\item{pi1}{nx1 vector with estimates for P(A=1 | X=x) evaluated at each x}

\item{mu0}{nx1 vector with estimates for E(Y |A=0, X=x) evaluated at each x}

\item{mu1}{nx1 vector with estimates for E(Y |A=1, X=x) evaluated at each x}

\item{upper}{boolean (default is upper=FALSE) if TRUE if values for 
E(g_u(eta)) are returned, otherwise those for E(g_l(eta)) are returned}
}
\value{
a length(y)x1 matrix containing the influence function values at 
each observed X of the parameter E(g(eta)).
}
\description{
\code{if_tau} returns the influence function values for the 
parameter E(g_l(eta)) (upper=FALSE) or E(g_u(eta)) (upper=TRUE) at each X.
}
\section{Details}{

As done in the paper, one can see that g(eta) for the lower bound is equal to
g(eta) for the upper bound minus delta * (ymax - ymin). Therefore the IFs for
E(g(eta)) follows the same relation. They are keep separated just for code
clarity.
}

\seealso{
\code{\link{if_gamma}}, \code{\link{get_ifvals}}.
}
