% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_prob}
\alias{truncate_prob}
\title{truncate_prob}
\usage{
truncate_prob(x, tol = 0.05)
}
\arguments{
\item{x}{a vector.}

\item{tol}{scalar used to make each observation in x be in [tol, 1-tol].}
}
\value{
a new vector of length(x) truncated.
}
\description{
\code{truncate_prob} truncates a vector of observations in [0, 1] so that 
each observation is in [tol, 1-tol].
}
